apihelper/requirements.txt
Flask==2.1.1
Flask-Cors==3.0.10
flask-restx==0.5.1
python-dotenv==0.20.0
gunicorn==20.1.0
requests